requirements-dev.txt
# Development dependencies for UEIL Data Store
# Core backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
minio>=7.2.0
psycopg2-binary>=2.9.7
gitpython>=3.1.40
PyYAML>=6.0.1
requests>=2.31.0
pydantic-settings>=2.0.3

# Frontend dependencies
streamlit>=1.28.0
pandas>=2.1.0

# Development tools
black>=23.0.0
isort>=5.12.0
pylint>=3.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Testing dependencies
requests>=2.31.0
pyyaml>=6.0.1

# Client library dependencies (if needed)
# These will be installed from the clients/python directory
